Some code review changes
* Add prettier as dependency
* Add patch-package and use it to patch patch-package and react-table
* Update docker config to work with the patches.
* Address some misc comments (file names etc)

Patching react-table is surprisingly annoying: because we need to patch
the dist files, we need to recompile them to patch them, which means
going to the react-table directory, running yarn install, grabbing
two config files that the developer didn't bother distributing, and
running yarn build. Then deleting the yarn.lock file and react-table's
node_modules directory to make the patch manageable. I included the two
missing configs in the patch so we won't have to worry about them, but
you will need to do the whole yarn install and yarn build rigamarole to
patch it again (which takes a million years because this package has a
million dependencies because of course it does).