Associate puzzle tags with hunts
This is needed to sandbox hunts from each other. Before, all the tags
from hunt 1 would appear in the add tags form for hunt 2, and I think
there would be bad interactions if we ever had eg two instances of the
same hunt on the site or a meta name in one hunt that was also a tag
name in a second hunt.