Switch tags from taggit to a simple ManyToMany model
We didn't really need any of the functionality that taggit provided and
it was pretty hard to work with: trying to write the migration to add
hunt ids to tags was causing me a lot of pain and I've also had
problems trying to use it with django rest framework.

The change in the models themselves is fairly minimal but the migration
to switch from taggit to this new m2m was kind of complicated. I used a
couple SQL queries to save the existing tag relationships to a temp
table and then to dump that table into the m2m relationship that django
created. Then in python I assign hunt_ids to tags based on the puzzles
they're associated with and create new tags when a tag is used in
multiple hunts. The end result is that the transition should preserve
everyone's local data.