Improve performance of main /puzzles API call (#324)
I installed django debug toolbar locally and used it to profile the API.
I found that on last year's hunt we were making 532 queries to load all
the puzzles due to bad prefetching:

* We were prefetching metas but not feeders
* We were sorting the tags which invalidated the prefetch for them
* We were filtering guesses to only show correct ones which invalidated
  the prefetch for them as well.

To fix, this adds prefetching for feeders, moves the tag sorting into
the tag model, and does the correct answer filtering inside the
prefetch. This brings it down to only 7 queries for last year's hunt
which is much more reasonable.