Try revamping the edit tags modal
Instead of a display of tags at the top that you can click to add, and a
list of all your tags at the bottom that you can use to remove, try just
toggling each tag on/off.

Played around a little while with the visuals and this is the best I had
so far but it's still a little ugly imo.

Thinking of using a UI like this to implement #68