Use sheet redirect endpoint on the frontend
Makes /puzzles responses almost 25% shorter. Hard to say precisely how
it affects the response time (too variable) but it appears to be a nice
improvement as well which is what I would expect. The only downside is
that opening a sheet is now a little slower since we have to do the
redirect.