Check list of whitelisted emails at authentication time (fixes #500)
Before, the list of whitelisted emails was only calculated once at start-up time, and every time you added a user, you would need to restart Cardboard before that user could log in.

Now we just query the list of emails added to the Google Drive as part of the authentication flow. So as soon as you add a user to a Google Drive, they should be able to login to Cardboard.

The Google Drive API to get the folder permissions is pretty fast, so it doesn't noticeably slow down the login. Also, according to https://support.google.com/a/answer/10445916?hl=en, the default Drive API quota is 10,000 calls every 100 seconds, which is plenty for us.

My implementation of `auth_allowed` is based off the default `auth_allowed` implementation here:
* https://github.com/python-social-auth/social-core/blob/911957bc2c0cf0f2e3419e1f1e7caaf25059dc6d/social_core/pipeline/social_auth.py#L12-L14
* https://github.com/python-social-auth/social-core/blob/911957bc2c0cf0f2e3419e1f1e7caaf25059dc6d/social_core/backends/base.py#L148-L159