Don't show meta tags by default for nested metas
We made a change to not show meta tags if there's only 1, which reduces a lot of clutter.
Feeder puzzles to a single meta don't have that meta's tag by default already.

However, nested meta puzzles would still show the tag for their parent meta. I think those
are unnecessary as well (it's conveyed by the table indentaiton) so this commit hides them
for the case of a single parent meta. If there are multiple parent metas then the tags are
shown again.